pyproject.toml
[project]
name = "second-brain"
version = "0.2.0"
description = "Automatisierte Second Brain Organisation mit Fabric AI + Obsidian"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Anton Feldmann", email = "anton@feldmann.dev"}
]
keywords = ["obsidian", "fabric", "second-brain", "pkm", "ai", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "python-frontmatter>=1.0.0",
    "PyYAML>=6.0",
]

[project.scripts]
# Haupt-Organisationstools
organize = "fabric_second_brain.organize:main"
init-vault = "fabric_second_brain.init_vault:main"
cleanup-vault = "fabric_second_brain.cleanup_vault:main"

# Daily Notes
create-dailies = "fabric_second_brain.create_dailies:main"

# YouTube Workflow
youtube = "fabric_second_brain.youtube_workflow:main"

# Generators
moc-builder = "fabric_second_brain.generators.moc_builder:main"
people-extractor = "fabric_second_brain.generators.people_extractor:main"
project-extractor = "fabric_second_brain.generators.project_extractor:main"
cluster-map = "fabric_second_brain.generators.cluster_map:main"

# Auto-Tools
auto-yaml-header = "fabric_second_brain.auto_yaml_header:main"
auto-tags = "fabric_second_brain.auto_tags:main"

[project.urls]
Homepage = "https://github.com/afeldman/fabric-second-brain"
Repository = "https://github.com/afeldman/fabric-second-brain"
Issues = "https://github.com/afeldman/fabric-second-brain/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["src/fabric_second_brain"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
